requirements.txt
echopype==0.8.1
functions==0.7.0
matplotlib==3.8.2
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.8.1.78
pandas==2.2.2
pyserial==3.5
PyYAML==6.0.1
PyYAML==6.0.1
ruamel.base==1.0.0
scipy==1.13.1
tqdm==4.64.0
xarray==2023.12.0
